pyproject.toml
[project]
name = "social-collector"
version = "0.1.0"
description = "Social media data collector"
requires-python = ">=3.9"
dependencies = [
    "tweety-ns",
    "aiohttp",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
